+++ packages/impit-client/src/index.ts
@@ -41,6 +41,7 @@ export class ImpitHttpClient implements BaseHttpClient {
             const reader = body.getReader();
             const buffer = new Uint8Array();
 
+            // eslint-disable-next-line no-constant-condition
             while (true) {
                 const { done, value } = await reader.read();