+++ packages/playwright-crawler/src/internals/utils/playwright-utils.ts
@@ -23,10 +23,10 @@ import vm from 'vm';
 
 import { LruCache } from '@apify/datastructures';
 import log_ from '@apify/log';
-import type { Request } from '@crawlee/browser';
-import { validators, KeyValueStore, RequestState, Configuration } from '@crawlee/browser';
+import type { Request, Session } from '@crawlee/browser';
+import { SessionError, validators, KeyValueStore, RequestState, Configuration } from '@crawlee/browser';
 import type { BatchAddRequestsResult } from '@crawlee/types';
-import { type CheerioRoot, type Dictionary, expandShadowRoots } from '@crawlee/utils';
+import { type CheerioRoot, type Dictionary, expandShadowRoots, sleep } from '@crawlee/utils';
 import * as cheerio from 'cheerio';
 import { getInjectableScript as getCookieClosingScript } from 'idcac-playwright';
 import ow from 'ow';
@@ -644,6 +644,145 @@ export async function closeCookieModals(page: Page): Promise<void> {
     await page.evaluate(getCookieClosingScript());
 }
 
+interface HandleCloudflareChallengeOptions {
+    /** Logging defaults to the `debug` level, use this flag to log to `info` level instead. */
+    verbose?: boolean;
+    /** How long should we wait after the challenge is completed for the final page to load. */
+    sleepSecs?: number;
+    /** Allows overriding the checkbox clicking. The `boundingBox` gives you approximate coordinates of the checkbox, use this if you need to adjust the click position. */
+    clickCallback?: (page: Page, boundingBox: { x: number; y: number }) => Promise<void>;
+    /** Allows overriding the detection of Cloudflare "challenge page". */
+    isChallengeCallback?: (page: Page) => Promise<boolean>;
+    /** Allows overriding the detection of Cloudflare "blocked page". */
+    isBlockedCallback?: (page: Page) => Promise<boolean>;
+}
+
+/**
+ * This helper tries to solve the Cloudflare challenge automatically by clicking on the checkbox.
+ * It will try to detect the Cloudflare page, click on the checkbox, and wait for 10 seconds (configurable
+ * via `sleepSecs` option) for the page to load. Use this in the `postNavigationHooks`, a failures will
+ * result in a SessionError which will be automatically retried, so only successful requests will get
+ * into the `requestHandler`.
+ *
+ * Works best with camoufox.
+ *
+ * **Example usage**
+ * ```ts
+ * postNavigationHooks: [
+ *     async ({ handleCloudflareChallenge }) => {
+ *         await handleCloudflareChallenge();
+ *     },
+ * ],
+ * ```
+ *
+ * @param page Playwright [`Page`](https://playwright.dev/docs/api/class-page) object
+ * @param url current URL for request identification, only used for logging
+ * @param [session] current session object
+ * @param [options]
+ */
+async function handleCloudflareChallenge(
+    page: Page,
+    url: string,
+    session?: Session,
+    options: HandleCloudflareChallengeOptions = {},
+): Promise<void> {
+    // eslint-disable-next-line dot-notation
+    const blockedStatusCodes = session?.['sessionPool']['blockedStatusCodes'] as number[];
+
+    // Cloudflare pages are 403, which are blocked by default
+    if (blockedStatusCodes?.includes(403)) {
+        const idx = blockedStatusCodes.indexOf(403);
+        blockedStatusCodes.splice(idx, 1);
+    }
+
+    options.isBlockedCallback ??= async () => {
+        const isBlocked = await page.evaluate(() => {
+            return document.querySelector('h1')?.textContent?.trim().includes('Sorry, you have been blocked');
+        });
+        return !!isBlocked;
+    };
+
+    options.isChallengeCallback ??= async () => {
+        return await page.evaluate(async () => {
+            return !!document.querySelector('.footer > .footer-inner > .diagnostic-wrapper > .ray-id');
+        });
+    };
+
+    const retryBlocked = async () => {
+        const isBlocked = await options.isBlockedCallback!(page).catch(() => false);
+
+        if (isBlocked) {
+            throw new SessionError(`Blocked by Cloudflare when processing ${url}`);
+        }
+    };
+
+    // check if we ended up on the CF challenge page
+    const isChallenge = async () => {
+        return options.isChallengeCallback!(page).catch(() => false);
+    };
+
+    if (!(await isChallenge())) {
+        await retryBlocked();
+        return;
+    }
+
+    const logLevel = options.verbose ? 'info' : 'debug';
+    log[logLevel](
+        `Detected Cloudflare challenge at ${url}, trying to solve it. This can take up to ${10 + (options.sleepSecs ?? 10)} seconds.`,
+    );
+
+    const bb = await page
+        .evaluate(() => {
+            const div = document.querySelector('.main-content div');
+            return div?.getBoundingClientRect();
+        })
+        .catch(() => undefined);
+
+    if (!bb) {
+        return;
+    }
+
+    const randomOffset = (range: number) => {
+        return Math.round(100 * range * Math.random()) / 100;
+    };
+
+    const x = bb.x + 30;
+    const y = bb.y + 25;
+
+    // try to click the checkbox every second
+    for (let i = 0; i < 10; i++) {
+        await sleep(1000);
+
+        // break early if we are no longer on the CF challenge page
+        if (!(await isChallenge())) {
+            break;
+        }
+
+        if (options.clickCallback) {
+            await options.clickCallback(page, { x, y });
+            continue;
+        }
+
+        // we can click on the text too, so X can be a bit larger
+        const xRandomized = x + randomOffset(10);
+        const yRandomized = y + randomOffset(10);
+
+        log[logLevel](`Trying to click on the Cloudflare checkbox at ${url}`, { x: xRandomized, y: yRandomized });
+        await page.mouse.click(xRandomized, yRandomized);
+
+        // sometimes the checkbox is lower (could be caused by a lag when rendering the logo)
+        await page.mouse.click(xRandomized, yRandomized + 35);
+    }
+
+    await sleep((options.sleepSecs ?? 10) * 1000);
+
+    if (await isChallenge()) {
+        throw new SessionError(`Blocked by Cloudflare when processing ${url}`);
+    }
+
+    await retryBlocked();
+}
+
 /** @internal */
 export interface PlaywrightContextUtils {
     /**
@@ -838,6 +977,28 @@ export interface PlaywrightContextUtils {
      * Tries to close cookie consent modals on the page. Based on the I Don't Care About Cookies browser extension.
      */
     closeCookieModals(): Promise<void>;
+
+    /**
+     * This helper tries to solve the Cloudflare challenge automatically by clicking on the checkbox.
+     * It will try to detect the Cloudflare page, click on the checkbox, and wait for 10 seconds (configurable
+     * via `sleepSecs` option) for the page to load. Use this in the `postNavigationHooks`, a failures will
+     * result in a SessionError which will be automatically retried, so only successful requests will get
+     * into the `requestHandler`.
+     *
+     * Works best with camoufox.
+     *
+     * **Example usage**
+     * ```ts
+     * postNavigationHooks: [
+     *     async ({ handleCloudflareChallenge }) => {
+     *         await handleCloudflareChallenge();
+     *     },
+     * ],
+     * ```
+     *
+     * @param [options]
+     */
+    handleCloudflareChallenge(options?: HandleCloudflareChallengeOptions): Promise<void>;
 }
 
 export function registerUtilsToContext(
@@ -881,6 +1042,9 @@ export function registerUtilsToContext(
         });
     context.compileScript = (scriptString: string, ctx?: Dictionary) => compileScript(scriptString, ctx);
     context.closeCookieModals = async () => closeCookieModals(context.page);
+    context.handleCloudflareChallenge = async (options?: HandleCloudflareChallengeOptions) => {
+        return handleCloudflareChallenge(context.page, context.request.url, context.session, options);
+    };
 }
 
 export { enqueueLinksByClickingElements };
@@ -898,4 +1062,5 @@ export const playwrightUtils = {
     compileScript,
     closeCookieModals,
     RenderingTypePredictor,
+    handleCloudflareChallenge,
 };